+++ pallets/ddc-payouts/src/benchmarking.rs
@@ -13,12 +13,17 @@ const CERE: u128 = 10000000000;
 
 fn create_dac_account<T: Config>() -> T::AccountId {
 	let dac_account = create_account::<T>("dac_account", 0, 0);
+	authorize_account::<T>(dac_account.clone());
 	T::ValidatorVisitor::setup_validators(vec![dac_account.clone()]);
 	dac_account
 }
 
 fn create_account<T: Config>(name: &'static str, idx: u32, seed: u32) -> T::AccountId {
 	account::<T::AccountId>(name, idx, seed)
+}
+
+fn authorize_account<T: Config>(account: T::AccountId) {
+	AuthorisedCaller::<T>::put(account);
 }
 
 fn endow_account<T: Config>(account: &T::AccountId, amount: u128) {
@@ -119,6 +124,14 @@ fn create_billing_report<T: Config>(params: BillingReportParams) {
 }
 
 benchmarks! {
+
+	set_authorised_caller {
+		 let dac_account = create_account::<T>("dac_account", 0, 0);
+
+	}: _(RawOrigin::Root, dac_account.clone())
+	verify {
+		assert_eq!(AuthorisedCaller::<T>::get(), Some(dac_account));
+	}
 
 	begin_billing_report {