+++ pallets/ddc-payouts/src/mock.rs
@@ -575,6 +575,7 @@ impl ExtBuilder {
 		let _payout_genesis = pallet_ddc_payouts::GenesisConfig::<Test> {
 			feeder_account: None,
 			debtor_customers: Default::default(),
+			authorised_caller: None,
 		}
 		.assimilate_storage(&mut storage);