+++ pallets/ddc-payouts/src/tests.rs
@@ -9,6 +9,29 @@ use sp_runtime::Perquintill;
 use super::{mock::*, *};
 
 #[test]
+fn set_authorised_caller_works() {
+	ExtBuilder.build_and_execute(|| {
+		System::set_block_number(1);
+
+		let root_account = 1u128;
+		let dac_account = 2u128;
+
+		assert_noop!(
+			DdcPayouts::set_authorised_caller(RuntimeOrigin::signed(root_account), dac_account),
+			BadOrigin
+		);
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
+
+		System::assert_last_event(
+			Event::AuthorisedCaller { authorised_caller: dac_account }.into(),
+		);
+
+		assert_eq!(DdcPayouts::authorised_caller().unwrap(), dac_account);
+	})
+}
+
+#[test]
 fn begin_billing_report_fails_for_unauthorised() {
 	ExtBuilder.build_and_execute(|| {
 		let root_account = 1u128;
@@ -21,6 +44,8 @@ fn begin_billing_report_fails_for_unauthorised() {
 		let start_era: i64 =
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_noop!(
 			DdcPayouts::begin_billing_report(
@@ -60,6 +85,8 @@ fn begin_billing_report_works() {
 		let start_era: i64 =
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
@@ -106,6 +133,8 @@ fn begin_charging_customers_fails_uninitialised() {
 			BadOrigin
 		);
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), DAC_ACCOUNT_ID));
+
 		assert_noop!(
 			DdcPayouts::begin_charging_customers(
 				RuntimeOrigin::signed(DAC_ACCOUNT_ID),
@@ -133,6 +162,8 @@ fn begin_charging_customers_works() {
 		let start_era: i64 =
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
@@ -202,6 +233,8 @@ fn send_charging_customers_batch_fails_uninitialised() {
 			),
 			BadOrigin
 		);
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_noop!(
 			DdcPayouts::send_charging_customers_batch(
@@ -420,6 +453,7 @@ fn send_charging_customers_batch_works() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -434,7 +468,7 @@ fn send_charging_customers_batch_works() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -502,7 +536,7 @@ fn send_charging_customers_batch_works() {
 			.into(),
 		);
 
-		assert_eq!(System::events().len(), 4 + 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 5 + 3 + 1); // 1 for Currency::transfer
 
 		// batch 2
 		let mut before_total_customer_charge = report.total_customer_charge;
@@ -655,6 +689,7 @@ fn end_charging_customers_works_small_usage_1_hour() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (1.0 * 1.0 * 3600.0) as i64; // 1 hour
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -669,7 +704,7 @@ fn end_charging_customers_works_small_usage_1_hour() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -873,6 +908,7 @@ fn send_charging_customers_batch_works_for_day() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (1.0 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -887,7 +923,7 @@ fn send_charging_customers_batch_works_for_day() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -955,7 +991,7 @@ fn send_charging_customers_batch_works_for_day() {
 			.into(),
 		);
 
-		assert_eq!(System::events().len(), 4 + 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 5 + 3 + 1); // 1 for Currency::transfer
 
 		// batch 2
 		let mut before_total_customer_charge = report.total_customer_charge;
@@ -1130,6 +1166,7 @@ fn send_charging_customers_batch_works_for_day_free_storage() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (1.0 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -1144,7 +1181,7 @@ fn send_charging_customers_batch_works_for_day_free_storage() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -1212,7 +1249,7 @@ fn send_charging_customers_batch_works_for_day_free_storage() {
 			.into(),
 		);
 
-		assert_eq!(System::events().len(), 4 + 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 5 + 3 + 1); // 1 for Currency::transfer
 
 		// batch 2
 		let mut before_total_customer_charge = report.total_customer_charge;
@@ -1387,6 +1424,7 @@ fn send_charging_customers_batch_works_for_day_free_stream() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (1.0 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -1401,7 +1439,7 @@ fn send_charging_customers_batch_works_for_day_free_stream() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -1469,7 +1507,7 @@ fn send_charging_customers_batch_works_for_day_free_stream() {
 			.into(),
 		);
 
-		assert_eq!(System::events().len(), 4 + 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 5 + 3 + 1); // 1 for Currency::transfer
 
 		// batch 2
 		let mut before_total_customer_charge = report.total_customer_charge;
@@ -1644,6 +1682,7 @@ fn send_charging_customers_batch_works_for_day_free_get() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (1.0 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -1658,7 +1697,7 @@ fn send_charging_customers_batch_works_for_day_free_get() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -1726,7 +1765,7 @@ fn send_charging_customers_batch_works_for_day_free_get() {
 			.into(),
 		);
 
-		assert_eq!(System::events().len(), 4 + 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 5 + 3 + 1); // 1 for Currency::transfer
 
 		// batch 2
 		let mut before_total_customer_charge = report.total_customer_charge;
@@ -1901,6 +1940,7 @@ fn send_charging_customers_batch_works_for_day_free_put() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (1.0 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -1915,7 +1955,7 @@ fn send_charging_customers_batch_works_for_day_free_put() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -1983,7 +2023,7 @@ fn send_charging_customers_batch_works_for_day_free_put() {
 			.into(),
 		);
 
-		assert_eq!(System::events().len(), 4 + 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 5 + 3 + 1); // 1 for Currency::transfer
 
 		// batch 2
 		let mut before_total_customer_charge = report.total_customer_charge;
@@ -2158,6 +2198,7 @@ fn send_charging_customers_batch_works_for_day_free_storage_stream() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (1.0 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -2172,7 +2213,7 @@ fn send_charging_customers_batch_works_for_day_free_storage_stream() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		assert_ok!(DdcPayouts::send_charging_customers_batch(
@@ -2240,7 +2281,7 @@ fn send_charging_customers_batch_works_for_day_free_storage_stream() {
 			.into(),
 		);
 
-		assert_eq!(System::events().len(), 4 + 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 5 + 3 + 1); // 1 for Currency::transfer
 
 		// batch 2
 		let mut before_total_customer_charge = report.total_customer_charge;
@@ -2384,6 +2425,7 @@ fn send_charging_customers_batch_works_zero_fees() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -2398,7 +2440,7 @@ fn send_charging_customers_batch_works_zero_fees() {
 			era,
 			max_batch_index,
 		));
-		assert_eq!(System::events().len(), 2);
+		assert_eq!(System::events().len(), 3);
 
 		// batch 1
 		let mut report = DdcPayouts::active_billing_reports(cluster_id, era).unwrap();
@@ -2468,6 +2510,8 @@ fn end_charging_customers_fails_uninitialised() {
 			DdcPayouts::end_charging_customers(RuntimeOrigin::root(), cluster_id, era,),
 			BadOrigin
 		);
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_noop!(
 			DdcPayouts::end_charging_customers(RuntimeOrigin::signed(dac_account), cluster_id, era,),
@@ -2541,6 +2585,8 @@ fn end_charging_customers_works() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
+
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -2581,7 +2627,7 @@ fn end_charging_customers_works() {
 
 		let mut balance = Balances::free_balance(DdcPayouts::account_id());
 		assert_eq!(balance - Balances::minimum_balance(), charge);
-		assert_eq!(System::events().len(), 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 4 + 1); // 1 for Currency::transfer
 
 		assert_ok!(DdcPayouts::end_charging_customers(
 			RuntimeOrigin::signed(dac_account),
@@ -2608,7 +2654,7 @@ fn end_charging_customers_works() {
 		);
 
 		let transfers = 3 + 3 + 3 * 3; // for Currency::transfer
-		assert_eq!(System::events().len(), 4 + 1 + 3 + transfers);
+		assert_eq!(System::events().len(), 8 + 1 + 3 + transfers);
 
 		let report_after = DdcPayouts::active_billing_reports(cluster_id, era).unwrap();
 		assert_eq!(report_after.state, PayoutState::CustomersChargedWithFees);
@@ -2619,31 +2665,63 @@ fn end_charging_customers_works() {
 			.left_from_one();
 
 		balance = Balances::free_balance(TREASURY_ACCOUNT_ID);
-		assert_eq!(balance, get_fees(&cluster_id).treasury_share * charge);
+		let mut expected_fees = get_fees(&cluster_id).treasury_share * charge;
+		assert_eq!(balance, expected_fees);
 
 		balance = Balances::free_balance(RESERVE_ACCOUNT_ID);
-		assert_eq!(balance, get_fees(&cluster_id).cluster_reserve_share * charge);
+		expected_fees = get_fees(&cluster_id).cluster_reserve_share * charge;
+		assert_eq!(balance, expected_fees);
 
 		balance = Balances::free_balance(VALIDATOR1_ACCOUNT_ID);
 		let mut ratio = Perquintill::from_rational(
 			VALIDATOR1_SCORE,
 			VALIDATOR1_SCORE + VALIDATOR2_SCORE + VALIDATOR3_SCORE,
 		);
-		assert_eq!(balance, get_fees(&cluster_id).validators_share * ratio * charge);
+		expected_fees = get_fees(&cluster_id).validators_share * ratio * charge;
+		assert_eq!(balance, expected_fees);
+		System::assert_has_event(
+			Event::ValidatorRewarded {
+				cluster_id,
+				era,
+				validator_id: VALIDATOR1_ACCOUNT_ID,
+				amount: expected_fees,
+			}
+			.into(),
+		);
 
 		balance = Balances::free_balance(VALIDATOR2_ACCOUNT_ID);
 		ratio = Perquintill::from_rational(
 			VALIDATOR2_SCORE,
 			VALIDATOR1_SCORE + VALIDATOR2_SCORE + VALIDATOR3_SCORE,
 		);
-		assert_eq!(balance, get_fees(&cluster_id).validators_share * ratio * charge);
+		expected_fees = get_fees(&cluster_id).validators_share * ratio * charge;
+		assert_eq!(balance, expected_fees);
+		System::assert_has_event(
+			Event::ValidatorRewarded {
+				cluster_id,
+				era,
+				validator_id: VALIDATOR2_ACCOUNT_ID,
+				amount: expected_fees,
+			}
+			.into(),
+		);
 
 		balance = Balances::free_balance(VALIDATOR3_ACCOUNT_ID);
 		ratio = Perquintill::from_rational(
 			VALIDATOR3_SCORE,
 			VALIDATOR1_SCORE + VALIDATOR2_SCORE + VALIDATOR3_SCORE,
 		);
-		assert_eq!(balance, get_fees(&cluster_id).validators_share * ratio * charge);
+		expected_fees = get_fees(&cluster_id).validators_share * ratio * charge;
+		assert_eq!(balance, expected_fees);
+		System::assert_has_event(
+			Event::ValidatorRewarded {
+				cluster_id,
+				era,
+				validator_id: VALIDATOR3_ACCOUNT_ID,
+				amount: expected_fees,
+			}
+			.into(),
+		);
 
 		assert_eq!(
 			report_after.total_customer_charge.transfer,
@@ -2690,6 +2768,8 @@ fn end_charging_customers_works_zero_fees() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
+
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -2730,7 +2810,7 @@ fn end_charging_customers_works_zero_fees() {
 
 		let mut balance = Balances::free_balance(DdcPayouts::account_id());
 		assert_eq!(balance - Balances::minimum_balance(), charge);
-		assert_eq!(System::events().len(), 3 + 1); // 1 for Currency::transfer
+		assert_eq!(System::events().len(), 4 + 1); // 1 for Currency::transfer
 
 		assert_ok!(DdcPayouts::end_charging_customers(
 			RuntimeOrigin::signed(dac_account),
@@ -2739,7 +2819,7 @@ fn end_charging_customers_works_zero_fees() {
 		));
 
 		System::assert_has_event(Event::ChargingFinished { cluster_id, era }.into());
-		assert_eq!(System::events().len(), 4 + 1);
+		assert_eq!(System::events().len(), 5 + 1);
 
 		let report_after = DdcPayouts::active_billing_reports(cluster_id, era).unwrap();
 		assert_eq!(report_after.state, PayoutState::CustomersChargedWithFees);
@@ -2831,6 +2911,8 @@ fn begin_rewarding_providers_fails_uninitialised() {
 			),
 			BadOrigin
 		);
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_noop!(
 			DdcPayouts::begin_rewarding_providers(
@@ -2943,6 +3025,8 @@ fn begin_rewarding_providers_works() {
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
+
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -3038,6 +3122,8 @@ fn send_rewarding_providers_batch_fails_uninitialised() {
 			),
 			BadOrigin
 		);
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_noop!(
 			DdcPayouts::send_rewarding_providers_batch(
@@ -3224,6 +3310,8 @@ fn send_rewarding_providers_batch_works() {
 		let start_era: i64 =
 			DateTime::<Utc>::from_naive_utc_and_offset(start_date.and_time(time), Utc).timestamp();
 		let end_era: i64 = start_era + (30.44 * 24.0 * 3600.0) as i64;
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
@@ -3556,6 +3644,7 @@ fn send_rewarding_providers_batch_100_nodes_small_usage_works() {
 			payers.push(user_batch.clone());
 		}
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -3833,6 +3922,7 @@ fn send_rewarding_providers_batch_100_nodes_large_usage_works() {
 			payers.push(user_batch.clone());
 		}
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -4109,6 +4199,7 @@ fn send_rewarding_providers_batch_100_nodes_small_large_usage_works() {
 			payers.push(user_batch.clone());
 		}
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -4344,6 +4435,7 @@ fn send_rewarding_providers_batch_100_nodes_random_usage_works() {
 			payers.push(user_batch.clone());
 		}
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -4530,6 +4622,8 @@ fn end_rewarding_providers_fails_uninitialised() {
 			BadOrigin
 		);
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
+
 		assert_noop!(
 			DdcPayouts::end_rewarding_providers(
 				RuntimeOrigin::signed(dac_account),
@@ -4697,6 +4791,8 @@ fn end_rewarding_providers_works() {
 		let payers = vec![(user1, bucket_id1, usage1)];
 		let payees = vec![(node1, node_usage1)];
 
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
+
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),
 			cluster_id,
@@ -4794,6 +4890,8 @@ fn end_billing_report_fails_uninitialised() {
 			DdcPayouts::end_billing_report(RuntimeOrigin::root(), cluster_id, era,),
 			BadOrigin
 		);
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_noop!(
 			DdcPayouts::end_billing_report(RuntimeOrigin::signed(dac_account), cluster_id, era,),
@@ -4929,6 +5027,8 @@ fn end_billing_report_works() {
 		let bucket_id1 = 1;
 		let payers = vec![(user1, bucket_id1, CustomerUsage::default())];
 		let payees = vec![(node1, NodeUsage::default())];
+
+		assert_ok!(DdcPayouts::set_authorised_caller(RuntimeOrigin::root(), dac_account));
 
 		assert_ok!(DdcPayouts::begin_billing_report(
 			RuntimeOrigin::signed(dac_account),